[BEAM-13052] Restructure pubsublite folder to move non-user interface classes to internal folder
PubsubLiteIO is composed of quite a few classes that are not part of the user interface for this module. Adding support for x-lang transforms will make this worse.

Move everything that is not part of the intended user interface to the `internal` folder.
